server/requirements.txt
fastapi==0.105.0
uvicorn[standard]==0.24.0
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
pydantic==2.5.2
pydantic-settings==2.1.0
prometheus-client==0.17.1
prometheus-fastapi-instrumentator==6.0.0
argon2-cffi==23.1.0
httpx==0.25.1
apscheduler==3.10.4
prophet==1.1.5
pandas==2.1.3
python-dotenv==1.0.0
ruff==0.1.12
pytest==7.4.4
pytest-asyncio==0.21.1
alembic==1.12.1
aiosqlite==0.20.0
